Don't assume that the referenceName is not of the "chr_" format
Apparently the variantsGeneSearchHelper code assumes, because the GSC files usually did it this way, that the variantSets would actually have their references in the format "#" instead of "chr#", so it replaces the "chr" by default. However, this isn't correct for the INSPIRE dataset. Jimmy is going to look into whether or not any current files at GSC really do use the "#" format still.